+++ src/main/java/com/arom/polisee/domain/user/Role.java
@@ -1,0 +1,4 @@
+package com.arom.polisee.domain.user;
+public enum Role {
+   ROLE_ADMIN, ROLE_USER
+}
