--- src/main/java/com/arom/polisee/global/login/entity/Role.java
@@ -1,6 +1,0 @@
-package com.arom.polisee.global.login.entity;
-
-
-public enum Role {
-   ROLE_ADMIN, ROLE_USER
-}
